Use int64 for pr number (#1478)

+++ src/drivers/bitbucket.go
@@ -19,7 +19,7 @@ func (d *bitbucketCodeHostingDriver) CanBeUsed(driverType string) bool {
 	return driverType == "bitbucket" || d.hostname == "bitbucket.org"
 }
 
-func (d *bitbucketCodeHostingDriver) CanMergePullRequest(branch, parentBranch string) (canMerge bool, defaultCommitMessage string, pullRequestNumber int, err error) {
+func (d *bitbucketCodeHostingDriver) CanMergePullRequest(branch, parentBranch string) (canMerge bool, defaultCommitMessage string, pullRequestNumber int64, err error) {
 	return false, "", 0, nil
 }
